packages/mcp-server/pyproject.toml
[project]
name = "taskflow-mcp"
version = "0.1.0"
description = "TaskFlow MCP Server - Task management tools for AI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.22.0",
    "httpx>=0.28.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.0.0",
    "starlette>=0.45.0",
    "uvicorn>=0.34.0",
    "pyjwt[crypto]>=2.10.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.22.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py313"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/taskflow_mcp"]
